chore: move imageverify cel lib
