+++ pkg/cel/libs/imageverify/impl_test.go
@@ -1,4 +1,4 @@
-package imageverifierfunctions
+package imageverify
 
 import (
 	"testing"
