+++ pkg/cel/libs/imageverify/utils.go
@@ -1,4 +1,4 @@
-package imageverifierfunctions
+package imageverify
 
 import (
 	"github.com/kyverno/kyverno/api/policies.kyverno.io/v1alpha1"
